Convert filterCovariance input to double precision

+++ algorithms/timeClosestApproach/timeClosestApproach.cpp
@@ -51,8 +51,8 @@ void TimeClosestApproach::updateState(const uint64_t currentSimNanos) {
 
     Eigen::Vector3d r_BN_N = cArrayToEigenVector(navFilterMsgPayload.r_BN_N).cast<double>();
     Eigen::Vector3d v_BN_N = cArrayToEigenVector(navFilterMsgPayload.v_BN_N).cast<double>();
-    Eigen::Matrix<float, 6, 6> filterCovariance = Eigen::Matrix<float, 6, 6>::Zero();
-    filterCovariance = cArrayToEigenMatrix<float, 6, 6>(filterStatePayload.covar);
+    Eigen::Matrix<double, 6, 6> filterCovariance = Eigen::Matrix<double, 6, 6>::Zero();
+    filterCovariance = cArrayToEigenMatrix<double, 6, 6>(filterStatePayload.covar);
 
     TimeClosestApproachOutput out_algo = this->algorithm->update(r_BN_N, v_BN_N, filterCovariance);
     this->writeMessages(out_algo.tCA, out_algo.sigmaTca, currentSimNanos);
